Added functionality to Main()
Made command line access for the program. Added DatabaseUtils and access to the database. Everything works with the exception that it pulls the root list not the user specific ones. So if anyone else has todos that arent the same format it will throw an error.  That can be changed very easily , however. Got rid of ToDoItemStatus because it doesn't fit in the database very nicely. Got rid of formatted timestamp bullshit because of storage in the database and both ToDoItem and TimeStamp have toString() overrides.